+++ semcore/d3-chart/__tests__/index.test.tsx
@@ -2870,3 +2870,51 @@ describe('patterns rendering', () => {
     await expect(await snapshot(<Component />)).toMatchImageSnapshot(task);
   });
 });
+
+describe('Cigarette chart', () => {
+  beforeEach(cleanup);
+
+  test.concurrent('Should render Cigarette', async ({ task }) => {
+    const data = {
+      a: 3524,
+      b: 1344,
+      c: 6135,
+      d: 1456,
+      e: 1823,
+    };
+
+    const Component = () => {
+      return (
+        <Flex gap={8}>
+          <Chart.Cigarette
+            data={data}
+            plotWidth={280}
+            header={
+              <Text size={700} bold>
+                Total
+              </Text>
+            }
+            showLegend={true}
+            tooltipTitle='Some title for tooltip'
+            showTotalInTooltip={true}
+          />
+
+          <Chart.Cigarette
+            data={data}
+            plotHeight={200}
+            invertAxis={false}
+            header={
+              <Text size={700} bold>
+                Total
+              </Text>
+            }
+            showLegend={true}
+            tooltipViewType='single'
+          />
+        </Flex>
+      );
+    };
+
+    await expect(await snapshot(<Component />)).toMatchImageSnapshot(task);
+  });
+});