+++ semcore/date-picker/src/DateRangePicker.jsx
@@ -87,8 +87,9 @@ class DateRangePickerRoot extends RangePickerAbstract {
       onDisplayedPeriodChange,
       locale,
       disabled,
-      size,
+      disabledErrorText,
       getI18nText,
+      animationsDisabled,
     } = this.asProps;
     const value = visible
       ? [preselectedValue[0] || this.asProps.value[0], preselectedValue[1] || this.asProps.value[1]]
@@ -100,10 +101,11 @@ class DateRangePickerRoot extends RangePickerAbstract {
       onChange,
       onDisplayedPeriodChange,
       locale,
-      w: size === 'm' ? 270 : 300,
       disabledDates: disabled,
+      disabledErrorText,
       children: () => <InputTrigger.DateRange />,
       getI18nText,
+      animationsDisabled,
     };
   }
 }