+++ semcore/d3-chart/CHANGELOG.md
@@ -2,6 +2,36 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.38.0] - 2024-04-12
+
+### Added
+
+- `Cigarette` chart.
+
+## [3.37.4] - 2024-04-12
+
+### Fixed
+
+- Dot's size for line, aria and bar charts.
+
+## [3.37.3] - 2024-04-12
+
+### Fixed
+
+- background color for Bar chart from `--intergalactic-chart-palette-order-other-data` to `--intergalactic-chart-grid-bar-chart-base-bg`.
+
+## [3.37.2] - 2024-04-12
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/popper` [5.26.2 ~> 5.26.3]).
+
+## [3.37.1] - 2024-04-10
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/animation` [2.21.1 ~> 2.21.2], `@semcore/utils` [4.22.1 ~> 4.22.2], `@semcore/core` [2.19.1 ~> 2.19.2]).
+
 ## [3.37.0] - 2024-03-27
 
 ### Added