+++ semcore/popper/CHANGELOG.md
@@ -2,11 +2,43 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.26.3] - 2024-04-12
+
+### Fixed
+
+- In some cases popper mouseenter was unexpectedly ignored.
+
+## [5.26.2] - 2024-04-10
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/animation` [2.21.1 ~> 2.21.2], `@semcore/flex-box` [5.21.1 ~> 5.21.2], `@semcore/neighbor-location` [4.19.1 ~> 4.19.2], `@semcore/outside-click` [3.19.1 ~> 3.19.2], `@semcore/utils` [4.22.1 ~> 4.22.2], `@semcore/core` [2.19.1 ~> 2.19.2]).
+
+## [5.26.1] - 2024-09-04
+
+### Fixed
+
+- Holding mouse over a place where popper appear was preventing popper from disappear back.
+- Fast keyboard navigation over multiple closely placed poppers was causing multiple poppers to stay open.
+
 ## [5.26.0] - 2024-03-27
 
 ### Fixed
 
 - Select and similar components with disabled portal were not working properly being wrapped into `label` tag.
+
+## [5.25.2] - 2024-04-03
+
+### Fixed
+
+- Holding mouse over a place where popper appear was preventing popper from disappear back.
+- Fast keyboard navigation over multiple closely placed poppers was causing multiple poppers to stay open.
+
+## [5.25.1] - 2024-04-03
+
+### Fixed
+
+- Select with `disabpledPortal` inside a label was opening second time after selecting the option.
 
 ## [5.25.0] - 2024-03-27