+++ semcore/utils/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.23.0] - 2024-04-12
+
+### Added
+
+- Api in internal mechanism for recursive search of child elements.
+
+## [4.22.2] - 2024-04-09
+
+### Fixed
+
+- Fixed internal mechanism of converting React nodes into text string of it's content.
+
 ## [4.22.1] - 2024-04-05
 
 ### Added